pyproject.toml
[project]
name = "telegram-notifications"
version = "0.1.0"
description = "Notifications through Telegram API."
readme = "README.md"
authors = [
    { name = "richardpordan", email = "105643799+richardpordan@users.noreply.github.com" }
]
requires-python = ">=3.13"

dependencies = [
    "httpx>=0.28.1",
    "typer>=0.24.1",
]

[project.scripts]
telegram-notifications = "telegram_notifications.cli:app"

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "prek>=0.3.8",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.8",
]
test = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "telegram-notifications",
]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "RET", # flake8-return
    "ARG", # unused arguments
    "PTH", # use pathlib
    "RUF", # Ruff-specific rules
    "PL",  # pylint rules
    "S",   # security
    "T",   # Type hints
    "D",   # Docstring style
    "COM", # Complexity checks
    "PYI", # stub file checks
]

ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = [
    "S101", # Asserts in tests
    "D103"  # No docstring in tests
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.uv.workspace]
members = [
    ".",
    ".",
    ".",
    ".",
    ".",
    ".",
]

[tool.uv.sources]
telegram-notifications = { workspace = true, editable = true }

[tool.pytest.ini_options]
asyncio_mode = "auto"
